+++ metalearners/drlearner.py
@@ -6,7 +6,7 @@ from collections.abc import Mapping, Sequence
 
 import numpy as np
 from joblib import Parallel, delayed
-from typing_extensions import Self
+from typing_extensions import Any, Self
 
 from metalearners._typing import (
     Features,
@@ -405,6 +405,10 @@ class DRLearner(_ConditionalAverageOutcomeMetaLearner):
             )
 
         return pseudo_outcome
+
+    @property
+    def init_args(self) -> dict[str, Any]:
+        return super().init_args | {"adaptive_clipping": self.adaptive_clipping}
 
     def _necessary_onnx_models(self) -> dict[str, list[_ScikitModel]]:
         return {