+++ tests/test_metalearner.py
@@ -1127,3 +1127,22 @@ def test_validate_outcome_different_classes(implementation, use_pandas, rng):
         ValueError, match="have seen different sets of classification outcomes."
     ):
         ml.fit(X, y, w)
+
+
+@pytest.mark.parametrize(
+    "implementation",
+    [TLearner, SLearner, XLearner, RLearner, DRLearner],
+)
+def test_init_args(implementation):
+    ml = implementation(
+        True,
+        2,
+        LogisticRegression,
+        LinearRegression,
+        LogisticRegression,
+    )
+    ml2 = implementation(**ml.init_args)
+
+    assert set(ml.__dict__.keys()) == set(ml2.__dict__.keys())
+    for key in ml.__dict__:
+        assert ml.__dict__[key] == ml2.__dict__[key]